Add dedicated header for syscall_hook.c
Fix the redundant usage like "extern ssize_t (*real_sys_write)(...);".
After including "syscall_hook.h", then usage of any real_sys_{}
functions will be available.
Main purpose of new header, a.k.a. "syscall_hook.h" is to globalize
all real_sys_{} functions and give syscall_hook a dedicated header.
